refactored the upload_download.py into three different scripts for fitting into the pipeline
